backend/services/medication/requirements.txt
fastapi[standard]~=0.115.12
sqlalchemy~=2.0.40
python-dotenv~=1.1.0
pydantic~=2.11.4
fastdbx~=0.1.0
python-multipart